==> 1_3_Split.ipynb <==
# Package

import pandas as pd

# Importer les données stationnarisées

df = pd.read_csv("df.csv", index_col="date")
df_stationary = pd.read_csv("df_stationary.csv", index_col="date")

df_stationary = df_stationary.rename(columns = {"S&P 500" : "S&P500"})
df_stationary

df_stationary.to_csv("df_stationary.csv")

df.head()

df.index = pd.to_datetime(df.index)
df_stationary.index = pd.to_datetime(df_stationary.index) 

# Split les données

end_time_train = "1989-12" # 

df_train = df[df.index <= end_time_train]
df_test = df[df.index > end_time_train]

df_stationary_train = df_stationary[df_stationary.index <= end_time_train]
df_stationary_test = df_stationary[df_stationary.index > end_time_train]

print(df_stationary_train.index[-1])
print(df_stationary_test.index[0])

print(df_train.index[-1])
print(df_test.index[0])

# Exporter les splits

df_train.to_csv("df_train.csv")
df_test.to_csv("df_test.csv")

df_stationary_train.to_csv("df_stationary_train.csv")
df_stationary_test.to_csv("df_stationary_test.csv")
