==> 1_5__Ajouter UNRATE_lags.ipynb <==
# Packages

import pandas as pd

# Importation 

df_stationary_test = pd.read_csv("df_stationary_test.csv", index_col=0)

df_stationary_test.head()

# Ajouter la variable lags

df_stationary_test["UNRATE_lag12"] = df_stationary_test["UNRATE"]

# Exporter

df_stationary_test.to_csv("df_stationary_test.csv")
